refactor(grid): rename grid_helper module
